+++ abquant/monitor/logger.py
@@ -9,11 +9,11 @@ FORMAT = "%(message)s"
 
 
 class Logger:
-    def __init__(self, name="abquant", disable_logger=False):
+    def __init__(self, name="abquant", log_path="./logs/", strategy=None, disable_logger=False):
         self._logger = logging.getLogger(name)
-        if not disable_logger:
-            self._logger.setLevel(LOG_LEVEL)
-            self.config_logger()
+        self.config_logger(log_path, disable_logger)
+        self._logger_struct = logging.getLogger('monitor_struct_9527')
+        self.config_logger_struct(log_path, strategy)
 
     def get_logger(self):
         return self._logger
@@ -21,24 +21,23 @@ class Logger:
     def get_formatter(self):
         return logging.Formatter(FORMAT)
 
-    def get_handler(self, htype, log_path):
+    def get_handler(self, htype, log_file = None, level = LOG_LEVEL):
         handler = None
         if htype == 'stdout':
             handler = logging.StreamHandler()
         elif htype == 'file':
-            handler = TimedRotatingFileHandler(log_path + "abquant.log", when="D", encoding="UTF-8", backupCount=30)
-        handler.setLevel(LOG_LEVEL)
+            handler = TimedRotatingFileHandler(log_file, when="D", encoding="UTF-8", backupCount=30)
+        else:
+            return handler
+        handler.setLevel(level)
         handler.setFormatter(self.get_formatter())
         return handler
 
-    def config_logger(self, log_path=''):
-        logger2 = self._logger
-        if logger2 is None:
-            logger2 = logging.getLogger('abquant')
-        if log_path is None or len(log_path) == 0:
+    def config_logger(self, log_path = None, disable_logger = False):
+        if not self._logger:
+            return
+        if not log_path:
             log_path = os.getcwd() + "/logs/"
-        elif log_path[-1:] != "/":
-            log_path += "/"
         fret = os.access(log_path, os.F_OK)
         if not fret:
             print("log_path not exist, create")
@@ -47,11 +46,28 @@ class Logger:
             except Exception as e:
                 print(e)
                 return
-        wret = os.access(log_path, os.W_OK)
-        if not wret:
-            print("log_path cannot write")
+        # do not check write privilege, just throw exception
+        # wret = os.access(log_path, os.W_OK)
+        # if not wret:
+        #     print("log_path cannot write")
+        #     return
+        if disable_logger:
             return
-        logger2.addHandler(self.get_handler('file', log_path))
+        self._logger.setLevel(LOG_LEVEL)
+        self._logger.addHandler(self.get_handler('file', os.path.join(log_path, 'abquant.log')))
+
+    def config_logger_struct(self, log_path = None, strategy: str = None):
+        if not self._logger_struct:
+            return
+        if not log_path:
+            log_path = os.getcwd() + "/logs/"
+        if not strategy:
+            filetrunk = 'abquant'
+        else:
+            filetrunk = strategy.replace(' ', '_')
+        self._logger_struct.setLevel(LOG_LEVEL)
+        self._logger_struct.addHandler(self.get_handler('file', os.path.join(log_path, filetrunk + '.struct')))
+        self._logger_struct.propagate = False
 
     def debug(self, msg, *args, **kwargs):
         self._logger.debug(msg, *args, **kwargs)
@@ -65,6 +81,9 @@ class Logger:
     def info(self, msg, *args, **kwargs):
         self._logger.info(msg, *args, **kwargs)
 
+    def info_struct(self, msg, *args, **kwargs):
+        self._logger_struct.info(msg, *args, **kwargs)
+
     def warn(self, msg, *args, **kwargs):
         self._logger.warning(msg, *args, **kwargs)
 
@@ -72,19 +91,23 @@ class Logger:
         self._logger.error(msg, *args, **kwargs)
 
     def print_log_format(self, data):
-        logger2 = self._logger
+        logger = self._logger
+        logger_struct = self._logger_struct
         if isinstance(data, (int, float)):
-            logger2.info(str(data))
+            logger.info(str(data))
         elif isinstance(data, (list, tuple, set)):
-            logger2.info(str(data))
+            logger.info(str(data))
         elif isinstance(data, dict):
             try:
                 strategy_name = data.get("strategy_name")
                 event_type = data.get("event_type")
                 event_time = data.get("event_time")
                 payload = data.get("payload")
+                if event_type and event_type == "struct":
+                    logger_struct.info(json.dumps(data))
+                    return
                 if strategy_name is None and event_type is None and payload is None:
-                    logger2.info(json.dumps(data))
+                    logger.info(json.dumps(data))
                     return
                 if event_time is not None:
                     date_array = datetime.fromtimestamp(event_time)
@@ -119,7 +142,7 @@ class Logger:
                             formatStr += f'Strategy Heartbeat'
                         else:
                             formatStr += str(ptype)
-                        logger2.debug(formatStr)
+                        logger.debug(formatStr)
                         return
                     elif event_type == 'log':
                         gateway_name = data.get("gateway_name")
@@ -138,10 +161,10 @@ class Logger:
                         formatStr += json.dumps(payload)
                 else:
                     formatStr += f'{event_type} '
-                logger2.info(formatStr)
+                logger.info(formatStr)
             except Exception as e:
-                logger2.error(e)
+                logger.error(e)
         else:
-            logger2.info(str(data))
+            logger.info(str(data))
         pass